Extract examples agent to module
